# docs/swagger.yaml
basePath: /api
definitions:
  dto.AcceptFriendshipRequest:
    properties:
      friend-id:
        type: string
      user-id:
        type: string
    required:
    - friend-id
    - user-id
    type: object
...
  dto.AddFriendRequest:
    properties:
      friend-login:
        type: string
      user-id:
        type: string
    required:
    - friend-login
    - user-id
    type: object
  dto.AddGameRequest:
    properties:
      game:
        type: string
      user-id:
        type: string
    required:
    - game
    - user-id
    type: object
  dto.CancelFriendshipRequest:
    properties:
      friend-id:
        type: string
      user-id:
        type: string
    required:
    - friend-id
    - user-id
    type: object
...
  dto.DeleteGameRequest:
    properties:
      game:
        type: string
      user-id:
        type: string
    required:
    - game
    - user-id
    type: object
...
  entities.Game:
    properties:
      name:
        type: string
      num_of_events:
        type: integer
      num_of_players:
        type: integer
      rating:
        type: number
    type: object
...
  entities.User:
    properties:
      avatar:
        type: string
      chat_id:
        type: string
      discord:
        type: string
      games:
        items:
          type: string
        type: array
      id:
        type: string
      login:
        type: string
      num_of_ratings:
        type: integer
      rating:
        type: number
      telegram:
        type: string
      total_rating:
        type: integer
    type: object
host: localhost:1111
info:
  contact: {}
  description: Ненавижу эту хуйню бля
  title: На жри уебок доки свои не подавись (недоделанная кстати хихиххихихих))))))
  version: "1.0"
paths:
...
  /friends:
    get:
      consumes:
      - application/json
      description: Retrieve the list of friends for a user
      parameters:
      - in: query
        name: amount
        required: true
        type: integer
      - in: query
        name: page
        required: true
        type: integer
      - in: query
        name: user-id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of friends
          schema:
            items:
              $ref: '#/definitions/entities.User'
            type: array
        "400":
          description: '{\"error\":\"string\"}'
          schema:
            type: object
        "408":
          description: '{\"error\":\"string\"}'
          schema:
            type: object
        "500":
          description: '{\"error\":\"string\"}'
          schema:
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get user's friends list
      tags:
      - friends
  /friends/accept:
    post:
      consumes:
      - application/json
      description: Accept a pending friend request from another user
      parameters:
      - description: Accept Friendship Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.AcceptFriendshipRequest'
      produces:
      - application/json
      responses:
        "200":
          description: '{\"message\":\"string\"}'
          schema:
            type: object
        "400":
          description: '{\"error\":\"string\"}'
          schema:
            type: object
        "408":
          description: '{\"error\":\"string\"}'
          schema:
            type: object
        "500":
          description: '{\"error\":\"string\"}'
          schema:
            type: object
      summary: Accept a friend request
      tags:
      - friends
  /friends/add:
    post:
      consumes:
      - application/json
      description: Send a friend request to another user
      parameters:
      - description: Add Friend Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.AddFriendRequest'
      produces:
      - application/json
      responses:
        "200":
          description: '{\"message\":\"string\"}'
          schema:
            type: object
        "400":
          description: '{\"error\":\"string\"}'
          schema:
            type: object
        "408":
          description: '{\"error\":\"string\"}'
          schema:
            type: object
        "500":
          description: '{\"error\":\"string\"}'
          schema:
            type: object
      security:
      - ApiKeyAuth: []
      summary: Add a new friend
      tags:
      - friends
  /friends/cancel:
    post:
      consumes:
      - application/json
      description: Remove a friend or cancel a pending friend request
      parameters:
      - description: Cancel Friendship Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CancelFriendshipRequest'
      produces:
      - application/json
      responses:
        "200":
          description: '{\"message\":\"string\"}'
          schema:
            type: object
        "400":
          description: '{\"error\":\"string\"}'
          schema:
            type: object
        "408":
          description: '{\"error\":\"string\"}'
          schema:
            type: object
        "500":
          description: '{\"error\":\"string\"}'
          schema:
            type: object
      security:
      - ApiKeyAuth: []
      summary: Cancel a friendship
      tags:
      - friends
  /friends/requests:
    get:
      consumes:
      - application/json
      description: Retrieve pending friend requests for a user
      parameters:
      - in: query
        name: amount
        required: true
        type: integer
      - in: query
        name: page
        required: true
        type: integer
      - in: query
        name: user-id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of friend requests
          schema:
            items:
              $ref: '#/definitions/entities.User'
            type: array
        "400":
          description: '{\"error\":\"string\"}'
          schema:
            type: object
        "408":
          description: '{\"error\":\"string\"}'
          schema:
            type: object
        "500":
          description: '{\"error\":\"string\"}'
          schema:
            type: object
      summary: Get friend requests
      tags:
      - friends
  /games:
    delete:
      consumes:
      - application/json
      description: Remove a game from user's collection
      parameters:
      - description: Delete Game Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.DeleteGameRequest'
      produces:
      - application/json
      responses:
        "200":
          description: '{\"message\":\"string\"}'
          schema:
            type: object
        "400":
          description: '{\"error\":\"string\"}'
          schema:
            type: object
        "408":
          description: '{\"error\":\"string\"}'
          schema:
            type: object
        "500":
          description: '{\"error\":\"string\"}'
          schema:
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete a game from user
      tags:
      - games
    get:
      consumes:
      - application/json
      description: Retrieve paginated list of user's games
      parameters:
      - in: query
        name: amount
        required: true
        type: integer
      - in: query
        name: page
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of games
          schema:
            items:
              $ref: '#/definitions/entities.Game'
            type: array
        "400":
          description: '{\"error\":\"string\"}'
          schema:
            type: object
        "408":
          description: '{\"error\":\"string\"}'
          schema:
            type: object
        "500":
          description: '{\"error\":\"string\"}'
          schema:
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get user's games
      tags:
      - games
    post:
      consumes:
      - application/json
      description: Add a game to user's collection
      parameters:
      - description: Add Game Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.AddGameRequest'
      produces:
      - application/json
      responses:
        "200":
          description: '{\"message\":\"string\"}'
          schema:
            type: object
        "400":
          description: '{\"error\":\"string\"}'
          schema:
            type: object
        "408":
          description: '{\"error\":\"string\"}'
          schema:
            type: object
        "500":
          description: '{\"error\":\"string\"}'
          schema:
            type: object
      security:
      - ApiKeyAuth: []
      summary: Add a game to user
      tags:
      - games
  /games/{game}:
    get:
      consumes:
      - application/json
      description: Get detailed information about specific game
      parameters:
      - description: Game title
        in: path
        name: game
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Game data
          schema:
            $ref: '#/definitions/entities.Game'
        "400":
          description: '{\"error\":\"string\"}'
          schema:
            type: object
        "408":
          description: '{\"error\":\"string\"}'
          schema:
            type: object
        "500":
          description: '{\"error\":\"string\"}'
          schema:
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get game details
      tags:
      - games
...
swagger: "2.0"